# swagger.yml
openapi: 3.0.3
info:
  title: Swagger social network - OpenAPI 3.0
  description: |

  termsOfService: http://swagger.io/terms/
  contact:
    email: seotrangbh24@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://trang-dev.ictu.vn:10093/api
paths:
...
  /file/img:
    post:
      tags:
        - File
      summary: "Upload file images"
      security:
        - JWT: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                img:
                  type: string
                  format: binary
                  description: "The image file"
              required:
                - img
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  filepath:
                    type: string
                    description: "Path to the uploaded image"
                    example: "/uploads/image.jpg"
  /file/files:
    post:
      tags:
        - File
      summary: "Upload file images"
      security:
        - JWT: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  format: binary
                  description: "The image file"
              required:
                - files
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  filepath:
                    type: array
                    description: "Path to the uploaded image"
                    example: ["/uploads/test.docx", "uploads/Book1.xlsx"]
  /file/deleteFile?filePath={filePath}:
    delete:
      tags:
        - File
      summary: "Delete file"
      security:
        - JWT: []
      responses:
        "200":
          description: "Successful operation"

  # address ---------------------------------------------------------
...